Expanding git-new-workdir to support different sparse checkout settings per working copy.
Support is added in the spirit of the original script, removing the info symlink and
replacing it by an rsync copy of the contents of the main working copy.

Note: the contents of other files in info are not maintained in sync.

Before using the tool, one must enable the sparse checkout mode using

git config core.sparsecheckout true